Resolving missing points at the last buffer
When src_process is called multiple times with a connected set of buffers, the last input data with the size of half_filter_chan_len is not processed. In other words, you will lose the last half_filter_chan_len points of data (which was 47 samples when I tried with SRC_SINC_MEDIUM_QUALITY). This is because the while loop on line 375 breaks when filter->out_gen equals filter->out_count. Logically it makes sense, except that it should further generate output (half_filter_chan_len points more) for the very last buffer, because the processing of the very first buffer was done for filter->out_count MINUS half_filter_chan_len. 

To resolve this issue, I am simply changing line 375 to while (1) and let the separate termination check do the work (lines 390 and 393).  And I am making sure, in my application, to allocate the data_out buffer big enough to absorb this in addition to the expected output buffer size.

The same logic should apply to other functions--sinc_stereo_vari_process, sinc_quad_vari_process, sinc_hex_vari_process, and sinc_multichan_vari_process, but I'm not using them and I can't check this issue in these functions. I would let others do it. But this change works well for my application, auxlab, so I'm using this modified version for it. Maybe there are other ways to resolve this issue; but I would like to see it resolved one way or another, so I can continue to use the "official" libsamplerate.